pyproject.toml
[tool.poetry]
name = "fml_doc_gen"
version = "3.0.4"
description = "doc string generator"
authors = ["Farhan Faisal", "Michael Suriawan", "Lukman Lateef"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
poetry = "^2.0.0"
myst-nb = "^1.1.2"
autoapi = "^2.0.1"
colorama = "^0.4.6"
openai = "^1.60.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
sphinx = "^8.1.3"
sphinx-autoapi = "^3.4.0"
sphinx-rtd-theme = "^3.0.2"
jupyter = "^1.1.1"
myst-nb = "^1.1.2"
furo = "^2024.8.6"

[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]                                                    # version location
branch = "main"                                      # branch to make releases of
changelog_file = "CHANGELOG.md"                      # changelog file
build_command = "pip install poetry && poetry build" # build dists 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
